Add boustrophedon_client example package
This package is a typical example of path planner client. It loads a
boundary (stored in boundary.txt), send it to the boustrophedon_server,
wait for the resulting path, and write it in a text file (path.txt).

Two scripts are provided to generate a 2D boundary (create_boundary.py),
and visualize the final path within this boundary (display_path.py).